requirements.txt
selenium==4.23.1
webdriver-manager==4.0.2
pymorphy3==2.0.2
pandas==2.2.2
requests==2.32.3
urllib3==2.2.2
python-dotenv
openpyxl
